MAINT: Move function (#9165)
* MAINT: Move function

* FIX: Flake
